+++ src/aks-preview/azext_aks_preview/tests/latest/test_agentpool_decorator.py
@@ -29,7 +29,6 @@ from azure.cli.command_modules.acs._consts import (
     CONST_DEFAULT_NODE_VM_SIZE,
 )
 from azext_aks_preview._consts import (
-    CONST_DEFAULT_AUTOMATIC_SKU_NODE_VM_SIZE,
     CONST_DEFAULT_WINDOWS_NODE_VM_SIZE,
 )
 from azure.cli.command_modules.acs.agentpool_decorator import AKSAgentPoolParamDict
@@ -770,7 +769,7 @@ class AKSPreviewAgentPoolContextCommonTestCase(unittest.TestCase):
             DecoratorMode.CREATE,
             self.agentpool_decorator_mode,
         )
-        self.assertEqual(ctx_5.get_node_vm_size(), CONST_DEFAULT_AUTOMATIC_SKU_NODE_VM_SIZE)
+        self.assertEqual(ctx_5.get_node_vm_size(), "")
 
     def common_get_gateway_prefix_size(self):
         # default