+++ src/serviceconnector-passwordless/azext_serviceconnector_passwordless/_credential_free.py
@@ -2,7 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-# pylint: disable=no-member, too-many-lines, anomalous-backslash-in-string, redefined-outer-name, no-else-raise, attribute-defined-outside-init
+# pylint: disable=no-member, too-many-lines, anomalous-backslash-in-string, redefined-outer-name, no-else-raise, attribute-defined-outside-init,too-many-positional-arguments
 
 import struct
 import sys
@@ -989,7 +989,7 @@ class SourceHandler:
 
 
 def output_is_none(output):
-    return not output.stdout
+    return not output.result
 
 
 class LocalHandler(SourceHandler):