+++ src/aks-preview/azext_aks_preview/_params.py
@@ -129,6 +129,8 @@ from azext_aks_preview._consts import (
     CONST_APP_ROUTING_NONE_NGINX,
     CONST_TLS_MANAGEMENT_MANAGED,
     CONST_TLS_MANAGEMENT_NONE,
+    CONST_GPU_DRIVER_TYPE_CUDA,
+    CONST_GPU_DRIVER_TYPE_GRID,
 )
 from azext_aks_preview._validators import (
     validate_acr,
@@ -418,6 +420,11 @@ app_routing_nginx_configs = [
 tls_management_types = [
     CONST_TLS_MANAGEMENT_MANAGED,
     CONST_TLS_MANAGEMENT_NONE,
+]
+
+gpu_driver_types = [
+    CONST_GPU_DRIVER_TYPE_CUDA,
+    CONST_GPU_DRIVER_TYPE_GRID,
 ]
 
 
@@ -1589,6 +1596,11 @@ def load_arguments(self, _):
             help="space-separated tags: key[=value] [key[=value] ...].",
         )
         c.argument('skip_gpu_driver_install', action='store_true', is_preview=True)
+        c.argument(
+            "driver_type",
+            arg_type=get_enum_type(gpu_driver_types),
+            is_preview=True,
+        )
         # in creation scenario, use "localuser" as default
         c.argument(
             'ssh_access',