+++ src/netappfiles-preview/azext_netappfiles_preview/aaz/latest/netappfiles/volume/_show.py
@@ -17,12 +17,15 @@ from azure.cli.core.aaz import *
 )
 class Show(AAZCommand):
     """Get the details of the specified volume
+
+    :example: Returns the properties of the given ANF volume
+        az netappfiles volume show -g mygroup --account-name myaccname --pool-name mypoolname --name myvolname
     """
 
     _aaz_info = {
-        "version": "2022-11-01-preview",
+        "version": "2024-03-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.netapp/netappaccounts/{}/capacitypools/{}/volumes/{}", "2022-11-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.netapp/netappaccounts/{}/capacitypools/{}/volumes/{}", "2024-03-01-preview"],
         ]
     }
 
@@ -48,7 +51,7 @@ class Show(AAZCommand):
             required=True,
             id_part="name",
             fmt=AAZStrArgFormat(
-                pattern="^[a-zA-Z0-9][a-zA-Z0-9\-_]{0,127}$",
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9\\-_]{0,127}$",
             ),
         )
         _args_schema.pool_name = AAZStrArg(
@@ -57,7 +60,7 @@ class Show(AAZCommand):
             required=True,
             id_part="child_name_1",
             fmt=AAZStrArgFormat(
-                pattern="^[a-zA-Z0-9][a-zA-Z0-9\-_]{0,63}$",
+                pattern="^[a-zA-Z0-9][a-zA-Z0-9\\-_]{0,63}$",
                 max_length=64,
                 min_length=1,
             ),
@@ -71,7 +74,7 @@ class Show(AAZCommand):
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
-                pattern="^[a-zA-Z][a-zA-Z0-9\-_]{0,63}$",
+                pattern="^[a-zA-Z][a-zA-Z0-9\\-_]{0,63}$",
                 max_length=64,
                 min_length=1,
             ),
@@ -151,7 +154,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-11-01-preview",
+                    "api-version", "2024-03-01-preview",
                     required=True,
                 ),
             }
@@ -210,6 +213,9 @@ class Show(AAZCommand):
             _schema_on_200.zones = AAZListType()
 
             properties = cls._schema_on_200.properties
+            properties.accept_grow_capacity_pool_for_short_term_clone_split = AAZStrType(
+                serialized_name="acceptGrowCapacityPoolForShortTermCloneSplit",
+            )
             properties.actual_throughput_mibps = AAZFloatType(
                 serialized_name="actualThroughputMibps",
                 flags={"read_only": True},
@@ -235,6 +241,9 @@ class Show(AAZCommand):
             )
             properties.cool_access = AAZBoolType(
                 serialized_name="coolAccess",
+            )
+            properties.cool_access_retrieval_policy = AAZStrType(
+                serialized_name="coolAccessRetrievalPolicy",
             )
             properties.coolness_period = AAZIntType(
                 serialized_name="coolnessPeriod",
@@ -279,6 +288,11 @@ class Show(AAZCommand):
                 serialized_name="fileSystemId",
                 flags={"read_only": True},
             )
+            properties.inherited_size_in_bytes = AAZIntType(
+                serialized_name="inheritedSizeInBytes",
+                nullable=True,
+                flags={"read_only": True},
+            )
             properties.is_default_quota_enabled = AAZBoolType(
                 serialized_name="isDefaultQuotaEnabled",
             )
@@ -293,6 +307,9 @@ class Show(AAZCommand):
             )
             properties.key_vault_private_endpoint_resource_id = AAZStrType(
                 serialized_name="keyVaultPrivateEndpointResourceId",
+            )
+            properties.language = AAZStrType(
+                nullable=True,
             )
             properties.ldap_enabled = AAZBoolType(
                 serialized_name="ldapEnabled",
@@ -343,6 +360,7 @@ class Show(AAZCommand):
             )
             properties.smb_access_based_enumeration = AAZStrType(
                 serialized_name="smbAccessBasedEnumeration",
+                nullable=True,
             )
             properties.smb_continuously_available = AAZBoolType(
                 serialized_name="smbContinuouslyAvailable",
@@ -404,9 +422,6 @@ class Show(AAZCommand):
             )
 
             backup = cls._schema_on_200.properties.data_protection.backup
-            backup.backup_enabled = AAZBoolType(
-                serialized_name="backupEnabled",
-            )
             backup.backup_policy_id = AAZStrType(
                 serialized_name="backupPolicyId",
             )
@@ -418,8 +433,15 @@ class Show(AAZCommand):
             )
 
             replication = cls._schema_on_200.properties.data_protection.replication
+            replication.destination_replications = AAZListType(
+                serialized_name="destinationReplications",
+                flags={"read_only": True},
+            )
             replication.endpoint_type = AAZStrType(
                 serialized_name="endpointType",
+            )
+            replication.remote_path = AAZObjectType(
+                serialized_name="remotePath",
             )
             replication.remote_volume_region = AAZStrType(
                 serialized_name="remoteVolumeRegion",
@@ -430,9 +452,37 @@ class Show(AAZCommand):
             )
             replication.replication_id = AAZStrType(
                 serialized_name="replicationId",
+                flags={"read_only": True},
             )
             replication.replication_schedule = AAZStrType(
                 serialized_name="replicationSchedule",
+            )
+
+            destination_replications = cls._schema_on_200.properties.data_protection.replication.destination_replications
+            destination_replications.Element = AAZObjectType()
+
+            _element = cls._schema_on_200.properties.data_protection.replication.destination_replications.Element
+            _element.region = AAZStrType()
+            _element.replication_type = AAZStrType(
+                serialized_name="replicationType",
+            )
+            _element.resource_id = AAZStrType(
+                serialized_name="resourceId",
+            )
+            _element.zone = AAZStrType()
+
+            remote_path = cls._schema_on_200.properties.data_protection.replication.remote_path
+            remote_path.external_host_name = AAZStrType(
+                serialized_name="externalHostName",
+                flags={"required": True},
+            )
+            remote_path.server_name = AAZStrType(
+                serialized_name="serverName",
+                flags={"required": True},
+            )
+            remote_path.volume_name = AAZStrType(
+                serialized_name="volumeName",
+                flags={"required": True},
             )
 
             snapshot = cls._schema_on_200.properties.data_protection.snapshot