+++ src/nginx/azext_nginx/aaz/latest/nginx/deployment/configuration/_analyze.py
@@ -22,9 +22,9 @@ class Analyze(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-01-01-preview",
+        "version": "2024-06-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/nginx.nginxplus/nginxdeployments/{}/configurations/{}/analyze", "2024-01-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/nginx.nginxplus/nginxdeployments/{}/configurations/{}/analyze", "2024-06-01-preview"],
         ]
     }
 
@@ -201,7 +201,7 @@ class Analyze(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-01-01-preview",
+                    "api-version", "2024-06-01-preview",
                     required=True,
                 ),
             }